+++ src/rules/tailwind-sort-classes.ts
@@ -82,7 +82,7 @@ export const tailwindSortClasses: ESLintRule<Options> = {
               ...classes,
               unsortableClasses[1]
             ].join(""),
-            literal.openingQuote ?? '"'
+            literal.openingQuote ?? "`"
           );
 
           const fixedClasses =