+++ src/utils/matchers.test.ts
@@ -1,7 +1,5 @@
-import { equal } from "node:assert";
-import { describe, it } from "node:test";
-
 import { parse } from "espree";
+import { describe, expect, it } from "vitest";
 
 import {
   hasESNodeParentExtension,
@@ -46,9 +44,9 @@ describe("matchers", () => {
           return hasESNodeParentExtension(node) && isESObjectKey(node) && node.type === "Identifier" && node.name === "value";
         });
 
-        equal(getObjectPath(root[0]), "root");
-        equal(getObjectPath(nested[0]), "root.nested");
-        equal(getObjectPath(value[0]), "root.nested.value");
+        expect(getObjectPath(root[0])).toBe("root");
+        expect(getObjectPath(nested[0])).toBe("root.nested");
+        expect(getObjectPath(value[0])).toBe("root.nested.value");
 
       });
 
@@ -61,7 +59,7 @@ describe("matchers", () => {
         });
 
         const path = getObjectPath(value[0]);
-        equal(path, "root.nested.value");
+        expect(path).toBe("root.nested.value");
 
       });
 
@@ -89,9 +87,9 @@ describe("matchers", () => {
           return hasESNodeParentExtension(node) && isESObjectKey(node) && node.type === "Literal" && node.value === "deeply_nested_value";
         });
 
-        equal(getObjectPath(root[0]), `["root-key"]`);
-        equal(getObjectPath(nested[0]), `["root-key"]["1nested"]`);
-        equal(getObjectPath(value[0]), `["root-key"]["1nested"].deeply_nested_value`);
+        expect(getObjectPath(root[0])).toBe(`["root-key"]`);
+        expect(getObjectPath(nested[0])).toBe(`["root-key"]["1nested"]`);
+        expect(getObjectPath(value[0])).toBe(`["root-key"]["1nested"].deeply_nested_value`);
 
       });
     }
@@ -121,8 +119,8 @@ describe("matchers", () => {
           return hasESNodeParentExtension(node) && isESObjectKey(node) && node.type === "Identifier" && node.name === "value2";
         });
 
-        equal(getObjectPath(value1[0]), "root[0].value1");
-        equal(getObjectPath(value2[0]), "root[1].value2");
+        expect(getObjectPath(value1[0])).toBe("root[0].value1");
+        expect(getObjectPath(value2[0])).toBe("root[1].value2");
 
       });