+++ src/utils/quotes.test.ts
@@ -1,17 +1,16 @@
-import { equal } from "node:assert";
-import { describe, it } from "node:test";
+import { describe, expect, it } from "vitest";
 
 import { escapeNestedQuotes } from "readable-tailwind:utils:quotes.js";
 
 
 describe("escapeNestedQuotes", () => {
   it("should escape all nested quotes", () => {
-    equal(escapeNestedQuotes('content-[""]', "\""), 'content-[\\"\\"]');
-    equal(escapeNestedQuotes("content-['']", "'"), "content-[\\'\\']");
+    expect(escapeNestedQuotes('content-[""]', "\"")).toBe('content-[\\"\\"]');
+    expect(escapeNestedQuotes("content-['']", "'")).toBe("content-[\\'\\']");
   });
 
   it("should not escape quotes that are not nested", () => {
-    equal(escapeNestedQuotes('content-[""]', "'"), 'content-[""]');
-    equal(escapeNestedQuotes("content-['']", "\""), "content-['']");
+    expect(escapeNestedQuotes('content-[""]', "'")).toBe('content-[""]');
+    expect(escapeNestedQuotes("content-['']", "\"")).toBe("content-['']");
   });
 });