+++ src/utils/quotes.ts
@@ -2,5 +2,11 @@ import type { LiteralValueQuotes } from "readable-tailwind:types:ast.js";
 
 
 export function escapeNestedQuotes(content: string, surroundingQuotes: LiteralValueQuotes): string {
-  return content.replace(new RegExp(surroundingQuotes, "g"), `\\${surroundingQuotes}`);
+  const regex = surroundingQuotes === "'"
+    ? /(?<!\\)'/g
+    : surroundingQuotes === "\""
+      ? /(?<!\\)"/g
+      : /(?<!\\)`/g;
+
+  return content.replace(regex, `\\${surroundingQuotes}`);
 }