+++ vite.config.ts
@@ -1,0 +1,1 @@
+export { config as default } from "@schoero/configs/vite";